feat: add Helmholtz fonts
