+++ .gitignore
@@ -47,3 +47,6 @@ yarn-error.log*
 
 # typescript
 *.tsbuildinfo
+
+# Vim
+**/.*.sw[a-p]
