+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "frontend/assets/fonts/HelmholtzHalvar"]
+	path = frontend/assets/fonts/HelmholtzHalvar
+	url = git@github.com:hifis-net/RSD-halvar.git
