+++ frontend/pages/_document.tsx
@@ -23,10 +23,6 @@ export default class MyDocument extends Document {
         <Head>
           {/* Theme color for the browser, if it supports it, is REMOVED 2022-04-10 by Dusan */}
           {/* <meta name="theme-color" content={rsdMuiTheme.palette.primary.main} /> */}
-          {/* Roboto fonts */}
-          <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:100,300,400&display=swap"
-          />
-          <link href="https://fonts.googleapis.com/css2?family=Work+Sans:wght@700&display=swap" rel="stylesheet" />
           {/* add support for gracefull fallback for aos animations when js is disabled */}
           <noscript dangerouslySetInnerHTML={{__html: `
             <style type="text/css">
