+++ frontend/styles/global.css
@@ -4,6 +4,11 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
+@import url('~/assets/fonts/HelmholtzHalvar/HelmholtzHalvar.css');
+/* RSD default fonts:
+@import url('https://fonts.googleapis.com/css?family=Roboto:100,300,400&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Work+Sans:wght@700&display=swap');
+*/
 
 @tailwind base;
 @tailwind components;
