+++ frontend/styles/themeHelmholtzDark.js
@@ -27,7 +27,7 @@ const colors = {
   black:'#000',
   white: '#fff',
   background: '#fff',
-  paper: '#fff', /* Secondary Web Pale Blue */
+  paper: '#fff',
   grey: {
     50: '#fafafa',
     100: '#f5f5f5',
@@ -59,29 +59,6 @@ const colors = {
 //   xl: 1920
 // }
 
-const fonts={
-  // Note! If you're using custom (local or remote) font faces, those fonts must
-  // be loaded within styles/global.css
-  default: [
-    'Helmholtz Halvar Mittel Rg',
-    'Helvetica',
-    'arial',
-    'sans-serif'
-  ]
-}
-
-const muiTypography={
-  fontFamily: fonts.default.join(','),
-  // set default fontsize to 1rem for MUI-5
-  // fontSize:14,
-  fontWeightLight: 100,
-  fontWeightRegular: 300,
-  fontWeightMedium: 300,
-  fontWeightBold: 400,
-}
-
 module.exports={
-  colors,
-  fonts,
-  muiTypography
+  colors
 }
